refactor(shell): Phase 3b——OCR 帧处理与动作外置
ocrActions.js（560 行，18 个方法，3 段连续区间）：冻结帧取用/刷新/裁剪，
识别、复制、插入文档、收藏、生成网页链接、截图另存对话框。

project-overview.vue 5232 → 4675 行。

OCR 簇按"关注点 + 风险"切两刀。本次是低风险的一半：这 18 个方法不持有
跨方法的监听器实例态。采集入口、浮层生命周期、document 级 down/move/up
监听留到 Phase 3c——那部分持有 _ocrDoc* / _ocrGlobalBound，且改动要成对
检查绑定与解绑。

孤儿 import 只删本次改动造成的四个（ocrRecognize / createFile /
getApiBaseUrl / getSessionId，原本各只有 1 处引用且都在被搬走的方法里）。
另外发现 saveProjectVariable / getProjectVariables / logActivity /
inviteClient 在 HEAD 上就已无引用——属于既有死代码，本次不动，另行处理。

验证：
- 逐行比对：3 段区间共 560 行字节一致（按方法名定位，非硬编码行号）
- 方法名扫描：八个模块 80 个方法与 .vue 剩余 134 个方法零重名
- npm run build:h5 通过；npm run check:emits 通过
- npm run test:app-e2e 连续 2 次 30/30

注：OCR 链路无法在 app-e2e 浏览器目标里驱动（见 run.mjs J6.6 注释），
本次信心来自静态等价证明，不是端到端跑通。
